test(테스트 추가) 자동차 이동 테스트 추가
