+++ README.md
@@ -29,4 +29,4 @@
 - [x] 자동차를 전진 또는 멈춤 구현
 - [x] 자동차 상태화면 출력 구현
 - [x] 자동차 이름 입/출력 구현
-- [x] 레이싱 게임 우승자 출력 
+- [x] 레이싱 게임 우승자 출력
